refactor: 유사 도메인과 견적 도메인 구분
