+++ backend/src/main/java/softeer/wantcar/cartalog/similarity/controller/SimilarityController.java
@@ -1,4 +1,4 @@
-package softeer.wantcar.cartalog.estimate.controller;
+package softeer.wantcar.cartalog.similarity.controller;
 
 import io.swagger.annotations.*;
 import lombok.RequiredArgsConstructor;
@@ -6,9 +6,9 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import softeer.wantcar.cartalog.estimate.dto.SimilarEstimateCountResponseDto;
-import softeer.wantcar.cartalog.estimate.dto.SimilarEstimateResponseDto;
-import softeer.wantcar.cartalog.estimate.service.SimilarityService;
+import softeer.wantcar.cartalog.similarity.dto.SimilarEstimateCountResponseDto;
+import softeer.wantcar.cartalog.similarity.dto.SimilarEstimateResponseDto;
+import softeer.wantcar.cartalog.similarity.service.SimilarityService;
 
 @Api(tags = {"유사 견적 관련 API"})
 @RestController
