+++ backend/src/main/java/softeer/wantcar/cartalog/similarity/dto/SimilarEstimateResponseDto.java
@@ -1,4 +1,4 @@
-package softeer.wantcar.cartalog.estimate.dto;
+package softeer.wantcar.cartalog.similarity.dto;
 
 import lombok.Builder;
 import lombok.Getter;
