+++ backend/src/main/java/softeer/wantcar/cartalog/similarity/repository/SimilarityCommandRepository.java
@@ -1,6 +1,6 @@
-package softeer.wantcar.cartalog.estimate.repository;
+package softeer.wantcar.cartalog.similarity.repository;
 
-import softeer.wantcar.cartalog.estimate.repository.dto.SimilarityInfo;
+import softeer.wantcar.cartalog.similarity.repository.dto.SimilarityInfo;
 
 import java.util.List;
