+++ backend/src/main/java/softeer/wantcar/cartalog/similarity/repository/SimilarityQueryRepositoryImpl.java
@@ -1,4 +1,4 @@
-package softeer.wantcar.cartalog.estimate.repository;
+package softeer.wantcar.cartalog.similarity.repository;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.dao.DataAccessException;
@@ -7,12 +7,13 @@ import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
-import softeer.wantcar.cartalog.estimate.repository.dto.PendingHashTagMap;
-import softeer.wantcar.cartalog.estimate.repository.dto.SimilarityInfo;
+import softeer.wantcar.cartalog.similarity.repository.dto.PendingHashTagMap;
+import softeer.wantcar.cartalog.similarity.repository.dto.SimilarityInfo;
 import softeer.wantcar.cartalog.global.utils.RowMapperUtils;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.util.ArrayList;
 import java.util.List;
 
 @SuppressWarnings({"SqlResolve", "SqlNoDataSourceInspection"})
@@ -63,6 +64,9 @@ public class SimilarityQueryRepositoryImpl implements SimilarityQueryRepository 
 
     @Override
     public List<Long> findSimilarEstimateIds(List<Long> hashTagIndexes) {
+        if(hashTagIndexes.isEmpty()) {
+            return new ArrayList<>();
+        }
         SqlParameterSource parameters = new MapSqlParameterSource()
                 .addValue("hashTagIndexes", hashTagIndexes);
         return jdbcTemplate.queryForList(QueryString.findSimilarEstimateIds, parameters, Long.TYPE);
