+++ backend/src/main/java/softeer/wantcar/cartalog/similarity/service/SimilarityService.java
@@ -1,7 +1,7 @@
-package softeer.wantcar.cartalog.estimate.service;
+package softeer.wantcar.cartalog.similarity.service;
 
-import softeer.wantcar.cartalog.estimate.dto.SimilarEstimateCountResponseDto;
-import softeer.wantcar.cartalog.estimate.dto.SimilarEstimateResponseDto;
+import softeer.wantcar.cartalog.similarity.dto.SimilarEstimateCountResponseDto;
+import softeer.wantcar.cartalog.similarity.dto.SimilarEstimateResponseDto;
 
 public interface SimilarityService {
     SimilarEstimateResponseDto getSimilarEstimateInfo(Long estimateId, Long similarEstimateId);
