+++ backend/src/test/java/softeer/wantcar/cartalog/estimate/service/EstimateServiceImplTest.java
@@ -13,8 +13,10 @@ import softeer.wantcar.cartalog.estimate.dto.EstimateRequestDto;
 import softeer.wantcar.cartalog.estimate.dto.EstimateResponseDto;
 import softeer.wantcar.cartalog.estimate.repository.EstimateCommandRepository;
 import softeer.wantcar.cartalog.estimate.repository.EstimateQueryRepository;
-import softeer.wantcar.cartalog.estimate.repository.SimilarityCommandRepository;
-import softeer.wantcar.cartalog.estimate.repository.SimilarityQueryRepository;
+import softeer.wantcar.cartalog.estimate.service.EstimateService;
+import softeer.wantcar.cartalog.estimate.service.EstimateServiceImpl;
+import softeer.wantcar.cartalog.similarity.repository.SimilarityCommandRepository;
+import softeer.wantcar.cartalog.similarity.repository.SimilarityQueryRepository;
 import softeer.wantcar.cartalog.estimate.repository.dto.EstimateOptionIdListDto;
 import softeer.wantcar.cartalog.estimate.repository.dto.EstimateShareInfoDto;
 import softeer.wantcar.cartalog.model.repository.ModelOptionQueryRepository;
