+++ backend/src/test/java/softeer/wantcar/cartalog/similarity/repository/SimilarityQueryRepositoryTest.java
@@ -1,4 +1,4 @@
-package softeer.wantcar.cartalog.estimate.repository;
+package softeer.wantcar.cartalog.similarity.repository;
 
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.BeforeEach;
@@ -11,8 +11,10 @@ import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.test.context.jdbc.Sql;
 import org.springframework.transaction.annotation.Transactional;
-import softeer.wantcar.cartalog.estimate.repository.dto.PendingHashTagMap;
-import softeer.wantcar.cartalog.estimate.repository.dto.SimilarityInfo;
+import softeer.wantcar.cartalog.similarity.repository.dto.PendingHashTagMap;
+import softeer.wantcar.cartalog.similarity.repository.dto.SimilarityInfo;
+import softeer.wantcar.cartalog.similarity.repository.SimilarityQueryRepository;
+import softeer.wantcar.cartalog.similarity.repository.SimilarityQueryRepositoryImpl;
 
 import java.util.HashMap;
 import java.util.List;
